refactor: organize v2 recording and storage code
