+++ app/src/main/java/com/kooo/evcam/v2/permissions/AdbPermissionHelper.java
@@ -6,7 +6,6 @@ import android.os.Build;
 import android.os.Handler;
 import android.os.Looper;
 import android.util.Base64;
-import android.util.Log;
 
 import java.io.File;
 import java.io.FileInputStream;
